Print out the execution path along with the error message

+++ AcceptanceTests/AcceptanceTests.Common/AudioRecordings/AudioRecordingsManager.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Reflection;
 
 namespace AcceptanceTests.Common.AudioRecordings
 {
@@ -7,14 +8,23 @@ namespace AcceptanceTests.Common.AudioRecordings
     {
         public static string CreateNewAudioFile(string originalFileName, Guid hearingId, string path = "TestAudioFiles")
         {
-            var originalFile = $"{path}/{originalFileName}";
-            var newFileName = $"{path}/{hearingId}.mp4";
+            var originalFile = $"{path}\\{originalFileName}";
+            if (!File.Exists(originalFile))
+            {
+                throw new FileNotFoundException($"Unable to find audio file with path : {Assembly.GetExecutingAssembly().Location}\\{originalFile}. ");
+            }
+
+            var newFileName = $"{path}\\{hearingId}.mp4";
             File.Move($"{originalFile}", newFileName);
             return newFileName;
         }
 
         public static void RemoveLocalAudioFile(string filename)
         {
+            if (!File.Exists(filename))
+            {
+                throw new FileNotFoundException($"Unable to find audio file with path : {Assembly.GetExecutingAssembly().Location}\\{filename}");
+            }
             File.Delete(filename);
         }
     }
